+++ packages/Contracts/src/Http/Clients/Client.php
@@ -101,6 +101,11 @@ use Psr\Http\Message\UriInterface;
  * @method bool hasOption(string $name) Determine if a given option exists for the next request.
  * @method mixed getOption(string $name) Get a specific option for the next request.
  * @method array getOptions() Get all the options for the next request.
+ * @method Builder debug(callable|null $callback = null) Dumps the next outgoing request and received response.
+ * @method Builder dd(callable|null $callback = null) Dumps the next outgoing request and exists the running script.
+ * @method callable debugCallback() Returns the last assigned debugging callback. If no debugging method was assigned, method returns a callback that does not do anything.
+ * @method Builder log(callable|null $callback = null)  Log the next outgoing request and received response.
+ * @method callable logCallback() Returns the last assigned logging callback. If no method was assigned, method returns a callback that does not do anything.
  * @method self client() Get the http client.
  *
  * @method Builder select(string|string[] $field, string|null $resource = null) Select the fields to be returned.