+++ packages/Http/Clients/src/Providers/HttpClientServiceProvider.php
@@ -2,12 +2,8 @@
 
 namespace Aedart\Http\Clients\Providers;
 
-use Aedart\Contracts\Http\Clients\Manager as HttpClientsManager;
-use Aedart\Contracts\Http\Clients\Requests\Query\Grammars\Manager as HttpQueryGrammarsManager;
-use Aedart\Http\Clients\Manager as ClientsManager;
-use Aedart\Http\Clients\Requests\Query\Grammars\Manager as GrammarsManager;
-use Illuminate\Contracts\Support\DeferrableProvider;
-use Illuminate\Support\ServiceProvider;
+use Aedart\Http\Messages\Providers\HttpSerializationServiceProvider;
+use Illuminate\Support\AggregateServiceProvider;
 
 /**
  * Http Client Service Provider
@@ -15,28 +11,15 @@ use Illuminate\Support\ServiceProvider;
  * @author Alin Eugen Deac <aedart@gmail.com>
  * @package Aedart\Http\Clients\Providers
  */
-class HttpClientServiceProvider extends ServiceProvider implements DeferrableProvider
+class HttpClientServiceProvider extends AggregateServiceProvider
 {
-    public array $singletons = [
-        HttpClientsManager::class => ClientsManager::class,
-        HttpQueryGrammarsManager::class => GrammarsManager::class
+    /**
+     * Providers for the Http Clients package.
+     *
+     * @var string[]
+     */
+    protected $providers = [
+        ClientServiceProvider::class,
+        HttpSerializationServiceProvider::class
     ];
-
-    /**
-     * Bootstrap this service
-     */
-    public function boot()
-    {
-        $this->publishes([
-            __DIR__ . '/../../configs/http-clients.php' => config_path('http-clients.php')
-        ], 'config');
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function provides()
-    {
-        return [ HttpClientsManager::class, HttpQueryGrammarsManager::class ];
-    }
 }