+++ packages/Http/Clients/src/Requests/Builders/Guzzle/Pipes/AssignsPredefinedMiddleware.php
@@ -5,6 +5,8 @@ namespace Aedart\Http\Clients\Requests\Builders\Guzzle\Pipes;
 use Aedart\Contracts\Http\Clients\Middleware;
 use Aedart\Contracts\Http\Clients\Requests\Builder;
 use Aedart\Http\Clients\Middleware\AppliesResponseExpectations;
+use Aedart\Http\Clients\Middleware\RequestResponseDebugging;
+use Aedart\Http\Clients\Middleware\RequestResponseLogging;
 use Aedart\Http\Clients\Requests\Builders\ProcessedOptions;
 
 /**
@@ -22,7 +24,9 @@ class AssignsPredefinedMiddleware
      * @var string[]|Middleware[]
      */
     protected array $middleware = [
-        AppliesResponseExpectations::class
+        AppliesResponseExpectations::class,
+        RequestResponseLogging::class,
+        RequestResponseDebugging::class,
     ];
 
     /**