+++ packages/Http/Messages/src/Exceptions/UnableToSerializeResponse.php
@@ -1,0 +1,18 @@
+<?php
+
+
+namespace Aedart\Http\Messages\Exceptions;
+
+/**
+ * Unable To Serialize Response
+ *
+ * Should be thrown when unable to serialize a Http Response Message
+ *
+ * @see \Aedart\Http\Messages\Exceptions\HttpMessageSerializationException
+ *
+ * @author Alin Eugen Deac <aedart@gmail.com>
+ * @package Aedart\Http\Messages\Exceptions
+ */
+class UnableToSerializeResponse extends HttpMessageSerializationException
+{
+}