+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Added `QueueHandler`, a middleware processing component for Http Client Requests.
 * Added `AppliesResponseExpectations` middleware. Replacement for internal response expectations handling in Http `Client`.
 * Added `ResponseExpectation` component.
+* Added Http Message `Serializer` components, in `Aedart\Http\Messages` namespace.
+* Added `RequestResponseDebugging` Middleware for Http `Client`.
 * Support for loading [TOML](https://en.wikipedia.org/wiki/TOML) configuration files.
 * Added Http Request `Builder` aware of component.
 * PHP Compatibility check in Travis.
@@ -25,6 +27,7 @@
 * Changed `StatusCodesExpectation`, now inherits from `ResponseExpectation`. Some internal methods have been redesigned. This change should not affect your code, unless you have custom Http Request `Builder` implementation.
 * Changed `withExpectation()`, in Request `Builder`. Now accepts both a `callable` and a `ResponseExpectation` instance. This change should not affect your code, unless you have custom Http Request `Builder` implementation.
 * Changed Request `Builder` and Http `Client` interfaces and concrete implementations. Now offers methods for adding `Middleware`. This change only affects you if you have a custom Http `Client` or Request `Builder` implementation.
+* Changed Http `Client` and Request `Builder`, added debugging methods (`debug()`, `dd()`, `log()`...etc). This change only affects you if you have a custom Http `Client` or Request `Builder` implementation.
 * Removed `Aedart\Dto` (_was deprecated in `v4.x`_).
 * Removed `Aedart\ArrayDto` (_was deprecated in `v4.x`_).
 * Removed `Aedart\Console\CreateAwareOfCommand` (_was deprecated in `v4.x`_).
@@ -37,6 +40,7 @@
 
 * Added shortcut methods (_`getClient()` and `client()`_) for obtaining Http Client instance in `ProcessOptions`. 
 * Removed internal `applyExpectations()` method from `Expectations` concern, in Http Client `Builder`. Has been replaced by `AppliesResponseExpectations` middleware.
+* Changed `HttpClientServiceProvider`, now inherits from the `AggregateServiceProvider` and registers the `HttpSerializationServiceProvider` automatically. This eliminates setup of debugging components, for the Http `Client`.
 
 #### Fixed